Update MakeINI Script
Added a setup variables and constants section for information about the
environment in which this script will be run.

Added functionality for the script to understand where it resides and
reference files in folders relative to that path.

Added copy instructions for Iconic.Zip.dll which the executable requires
for proper functionality.

Added cleanup for the folder created by the header of the CSV file.